*move MergeGelu, MergeHswish, MergeMish, MergePrelu, MergeSwish and MergeConvolutionOrOtherAndActivation to separate source files.
